+++ mobile/lib/features/trade/trade_value_change_notifier.dart
@@ -1,12 +1,12 @@
 import 'package:flutter/material.dart';
+import 'package:get_10101/bridge_generated/bridge_definitions.dart' as bridge;
 import 'package:get_10101/common/application/channel_info_service.dart';
+import 'package:get_10101/common/application/event_service.dart';
 import 'package:get_10101/common/domain/model.dart';
+import 'package:get_10101/common/dummy_values.dart';
 import 'package:get_10101/features/trade/application/trade_values_service.dart';
 import 'package:get_10101/features/trade/domain/direction.dart';
 import 'package:get_10101/features/trade/domain/leverage.dart';
-import 'package:get_10101/bridge_generated/bridge_definitions.dart' as bridge;
-import 'package:get_10101/common/application/event_service.dart';
-import 'package:get_10101/common/dummy_values.dart';
 import 'package:get_10101/features/trade/domain/price.dart';
 import 'package:get_10101/features/trade/domain/trade_values.dart';
 
@@ -91,11 +91,11 @@ class TradeValuesChangeNotifier extends ChangeNotifier implements Subscriber {
     bool update = false;
 
     if (price.ask != _buyTradeValues.price) {
-      _buyTradeValues.updatePrice(price.ask);
+      _buyTradeValues.updatePriceAndQuantity(price.ask);
       update = true;
     }
     if (price.bid != _sellTradeValues.price) {
-      _sellTradeValues.updatePrice(price.bid);
+      _sellTradeValues.updatePriceAndQuantity(price.bid);
       update = true;
     }